CORE-1806 Add more details to Analyses Listing Details Drawer
Refactored details from AnalysisSubmissionLanding into
components/analyses/details/DetailsPanel,
which can also be used in the details Drawer.

Updated Drawer heading style so that long analysis names can wrap.
