+++ src/components/analyses/landing/AnalysisSubmissionLanding.js
@@ -21,17 +21,15 @@ import {
 } from "serviceFacades/support";
 import BatchResults from "./BatchResults";
 import DotMenuItems from "./DotMenuItems";
-import constants from "../../../constants";
 import AnalysisCommentDialog from "../AnalysisCommentDialog";
 import ids from "../ids";
 import RenameAnalysisDialog from "../RenameAnalysisDialog";
 import { canShare } from "../utils";
-import DataPathLink from "../../data/DataPathLink";
+import DetailsPanel from "../details/DetailsPanel";
 import InfoPanel from "../details/InfoPanel";
 import ParamsPanel from "../details/ParamsPanel";
 import AnalysisStatusIcon from "../AnalysisStatusIcon";
 
-import NavigationConstants from "common/NavigationConstants";
 import ShareWithSupportDialog from "components/analyses/ShareWithSupportDialog";
 import TerminateAnalysisDialog from "components/analyses/TerminateAnalysisDialog";
 import NotificationCategory from "components/models/NotificationCategory";
@@ -56,13 +54,8 @@ import analysisStatus from "components/models/analysisStatus";
 import Sharing from "components/sharing";
 import { formatSharedAnalyses } from "components/sharing/util";
 import ConfirmationDialog from "components/utils/ConfirmationDialog";
-import CopyLinkButton from "components/utils/CopyLinkButton";
-import { copyLinkToClipboardHandler } from "components/utils/copyLinkToClipboardHandler";
-import { copyStringToClipboard } from "components/utils/copyStringToClipboard";
 import { formatDate } from "components/utils/DateFormatter";
 import buildID from "components/utils/DebugIDUtil";
-import { getHost } from "components/utils/getHost";
-import GridLabelValue from "components/utils/GridLabelValue";
 import GridLoading from "components/utils/GridLoading";
 import { useConfig } from "contexts/config";
 import { useUserProfile } from "contexts/userProfile";
@@ -91,12 +84,9 @@ import BackButton from "components/utils/BackButton";
 import PendingTerminationDlg from "components/analyses/PendingTerminationDlg";
 import { openInteractiveUrl } from "../utils";
 
-const InfoGridValue = (props) => <Typography variant="body2" {...props} />;
-
 export default function AnalysisSubmissionLanding(props) {
     const { id, baseId, view, showErrorAnnouncer } = props;
     const { t } = useTranslation("analyses");
-    const { t: i18nCommon } = useTranslation("common");
     const theme = useTheme();
     const [userProfile] = useUserProfile();
     const [config] = useConfig();
@@ -480,71 +470,7 @@ export default function AnalysisSubmissionLanding(props) {
                     </Grid>
                 </Grid>
                 <Divider />
-                <Grid
-                    container
-                    spacing={3}
-                    style={{ marginTop: theme.spacing(1) }}
-                >
-                    <GridLabelValue label={t("app")}>
-                        <InfoGridValue>{analysis?.app_name}</InfoGridValue>
-                    </GridLabelValue>
-                    <GridLabelValue label={t("outputFolder")}>
-                        <div style={{ width: "100%" }}>
-                            <div style={{ float: "left" }}>
-                                {[
-                                    analysisStatus.SUBMITTED,
-                                    analysisStatus.RUNNING,
-                                ].includes(analysis?.status) && (
-                                    <Typography variant="body2">
-                                        {analysis?.resultfolderid}
-                                    </Typography>
-                                )}
-                                {[
-                                    analysisStatus.COMPLETED,
-                                    analysisStatus.FAILED,
-                                    analysisStatus.CANCELED,
-                                ].includes(analysis?.status) && (
-                                    <DataPathLink
-                                        id={baseId}
-                                        param_type="FolderInput"
-                                        path={analysis?.resultfolderid}
-                                    />
-                                )}
-                            </div>
-                            <div style={{ marginLeft: theme.spacing(0.25) }}>
-                                <CopyLinkButton
-                                    baseId={baseId}
-                                    onCopyLinkSelected={() => {
-                                        const link = `${getHost()}/${
-                                            NavigationConstants.DATA
-                                        }/${constants.DATA_STORE_STORAGE_ID}${
-                                            analysis?.resultfolderid
-                                        }`;
-                                        const copyPromise =
-                                            copyStringToClipboard(link);
-                                        copyLinkToClipboardHandler(
-                                            i18nCommon,
-                                            copyPromise
-                                        );
-                                    }}
-                                />
-                            </div>
-                        </div>
-                    </GridLabelValue>
-                    <GridLabelValue label={t("startDate")}>
-                        <InfoGridValue>
-                            {formatDate(analysis?.startdate)}
-                        </InfoGridValue>
-                    </GridLabelValue>
-                    <GridLabelValue label={t("endDate")}>
-                        <InfoGridValue>
-                            {formatDate(analysis?.enddate)}
-                        </InfoGridValue>
-                    </GridLabelValue>
-                    <GridLabelValue label={t("user")}>
-                        <InfoGridValue>{username}</InfoGridValue>
-                    </GridLabelValue>
-                </Grid>
+                <DetailsPanel baseId={baseId} analysis={analysis} />
                 <Accordion defaultExpanded={true}>
                     <AccordionSummary
                         expandIcon={<ExpandMore />}
